Remove "progress" region data collection
It turns out that the level region data contains the appropriate
information to determine which level each line of health and armor data
are pulled from, so that means the data collection from the progress
region was superfluous. Additionally, the name "progress" was a
minsnomer. That region can either contain the the text FINISHED, which
is appropriate for the name, or it can contain the name for the next
level, not ENTERING as I originally thought.

I any case, removing this part of the script will significantly speed up
it execution, though it will still remain painfully slow overall.